Added catalog page. Restructured src folder and all scenes/components for better reusability. Added loaders for fetchs
