+++ README.md
@@ -1,2 +1,2 @@
 # Front-end-courses-homework
-Don't even know why i am doing this anymore
+Homeworks from ITC front-end (fullstack) courses
