+++ delivery/package.json
@@ -6,6 +6,7 @@
     "react": "^16.3.2",
     "react-dom": "^16.3.2",
     "react-flexbox-grid": "^2.1.0",
+    "react-router-dom": "^4.2.2",
     "react-scripts": "1.1.4",
     "styled-components": "^3.2.6"
   },
