--- delivery/src/Buttons.js
@@ -1,22 +1,0 @@
-import styled from 'styled-components';
-
-const MainButton = styled.button`
-	padding: 12px 32px;
-    line-height: 1;
-	border: none;
-    border-radius: 34.5px;
-	opacity: .9;
-    color: #fff;
-	background-color: #a3d200;
-    cursor: pointer;
-	}
-	&:hover {
-		border-color: #b7e611;
-    	opacity: 1;
-	}
-	&:active {
-        border: none;
-    }
-`
-
-export default MainButton;
