--- delivery/src/Headers.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-const LeadHeader = styled.div`
-    font-family: "Pacifico", cursive;
-`
-
-export default LeadHeader;
