+++ delivery/src/Scenes/StoresCatalog/GreyLine.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const GreyLine = styled.hr`
+	border-color: #f6f6f6;
+	border-opacity: 0.8;
+	margin-top: 60px;
+`
+export default GreyLine;
