+++ delivery/src/index.js
@@ -1,8 +1,24 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Home from './Home.js';
+import {BrowserRouter, Switch, Route} from 'react-router-dom';
+
+import Home from './Scenes/Home.js';
+import StoresCatalog from './Scenes/StoresCatalog.js';
+
+class App extends React.Component {
+	render() {
+		return (
+				<Switch>
+					<Route exact path="/" component={Home}/>
+					<Route path="/catalog" component={StoresCatalog}/>
+				</Switch>
+			)
+	}
+}
 
 ReactDOM.render(
-    <Home />,
+    <BrowserRouter>
+    	<App/>
+	</BrowserRouter>,
     document.getElementById('root')
 );
